requirements.txt
kafka-python==2.0.2
datadog==0.44.0
python-snappy==0.6.1
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0


